source/lambda/active_learning_1p/requirements.txt
pillow>=7.0.0
numpy>=1.18.1
boto3>=1.13.1
